move 'unfold' tests to separate file
