+++ types/ramda/ramda-tests.ts
@@ -90,16 +90,6 @@ function addAll() {
 // Basic example
 R.useWith(addAll, [double, square]);
 
-(() => {
-    function f(n: number): false | [number, number] {
-        return n > 50 ? false : [-n, n + 10];
-    }
-
-    const a = R.unfold(f, 10); // => [-10, -20, -30, -40, -50]
-    const b = R.unfold(f); // => [-10, -20, -30, -40, -50]
-    const c = b(10);
-});
-
 /*********************
  * List category
  */
