+++ types/ramda/tsconfig.json
@@ -229,6 +229,7 @@
         "test/type-tests.ts",
         "test/unapply-tests.ts",
         "test/unary-tests.ts",
-        "test/uncurryN-tests.ts"
+        "test/uncurryN-tests.ts",
+        "test/unfold-tests.ts"
     ]
 }
